feat: migrate Gradle build configuration to root directory
The build and settings files in the `android/src/main/` directory were moved to `android/src/`.
A new `build.gradle` and `settings.gradle` were created in the `android/src/` directory.
The build configuration was updated to support kotlin and specify the project namespace.
The Kotlin version is defined in `build.gradle` and specified as 1.7.10.
The jvm target is specified as 1.8 in the `kotlinOptions`.